Significant re-engineering of Billboard and RenderBillboard.
Billboard object now contemplates holding either a URL or Base64 data.

RenderBillboard now determines how to load the image based on whether URL or Base64 data are non-null, and performs loading and scaling in its place. This has a significant bug-fix in it - under the old system, the 'lossless' data from the image was destroyed. Therefore, if the Billboard was previewed on a small screen, there would have been scaling issues when viewed full size in the Viewer.

Also implemented Billboard XML Reading